Mark symbols public instead of exporting.
Given the nature of this package, I think we should encourage explicit import of functionality in the packages that depend on GPUToolbox.

I originally didn't implement this, as Compat.jl adds a bunch of standard library dependencies to this no-dependency package, but then I found https://discourse.julialang.org/t/is-compat-jl-worth-it-for-the-public-keyword/119041/11.

It's not pretty, but it works.

Technically breaking, but all backends already explicity import the GPUToolbox functionality that they use, so it'll just be a matter of me adding "0.2" to the compat.